refactor: styled component 위치 이동
